Remove user-facing images in cluster config (no daemonset stuff)
